fix: unify rule config types
