+++ @commitlint/format/package.json
@@ -37,6 +37,7 @@
     "@commitlint/utils": "^8.3.4"
   },
   "dependencies": {
+    "@commitlint/types": "^8.3.5",
     "chalk": "^3.0.0"
   }
 }
