+++ @commitlint/types/src/index.ts
@@ -1,4 +1,5 @@
 export * from './ensure';
+export * from './format';
 export * from './is-ignored';
 export * from './rules';
 export * from './lint';
