+++ @commitlint/types/src/lint.ts
@@ -1,10 +1,11 @@
 import {IsIgnoredOptions} from './is-ignored';
-import {RuleConfigTuple, PluginRecords, RuleSeverity} from './load';
+import {PluginRecords} from './load';
 import {ParserOptions} from './parse';
+import {RuleConfigSeverity, RuleConfigTuple} from './rules';
 
 export type LintRuleConfig = Record<
 	string,
-	| Readonly<[RuleSeverity.Disabled]>
+	| Readonly<[RuleConfigSeverity.Disabled]>
 	| RuleConfigTuple<void>
 	| RuleConfigTuple<unknown>
 >;
@@ -35,7 +36,7 @@ export interface LintRuleOutcome {
 	/** If the commit is considered valid for the rule */
 	valid: boolean;
 	/** The "severity" of the rule (1 = warning, 2 = error) */
-	level: RuleSeverity;
+	level: RuleConfigSeverity;
 	/** The name of the rule */
 	name: string;
 	/** The message returned from the rule, if invalid */
